--- a/.github/workflows/ci.yaml
+++ b/.github/workflows/ci.yaml
@@ -5,6 +5,9 @@ on:
       - "main"
   push:
 
+env:
+  PRE_COMMIT_HOME: ".tox/pre-commit-home"
+
 jobs:
   lint:
     name: "Lint"
@@ -80,7 +83,7 @@ jobs:
           python-version: "3.11"
           cache: "pip"
           cache-dependency-path: |
-            .github/workflows/ci.yml
+            .github/workflows/ci.yaml
             pyproject.toml
             tox.ini
             week-number.txt
@@ -156,7 +159,7 @@ jobs:
           python-version: "${{ matrix.python-version }}"
           cache: "pip"
           cache-dependency-path: |
-            .github/workflows/ci.yml
+            .github/workflows/ci.yaml
             pyproject.toml
             tox.ini
             week-number.txt
@@ -173,7 +176,7 @@ jobs:
             week=${{ env.week-number }}
             os=${{ matrix.os.value }}
             python=${{ steps.setup-python.outputs.python-version }}
-            hash=${{ hashFiles('.github/workflows/ci.yml', 'pyproject.toml', 'tox.ini') }}
+            hash=${{ hashFiles('.github/workflows/ci.yaml', 'pyproject.toml', 'tox.ini') }}
 
       - name: "Identify virtual environment path"
         shell: "bash"