--- a/tox.ini
+++ b/tox.ini
@@ -19,6 +19,8 @@ commands = coverage erase
 
 
 [testenv]
+depends =
+    py{311, 310, 39, 38, 37}: coverage_erase
 extras =
     testing
     !minimum_flask: flask
@@ -28,6 +30,8 @@ commands = coverage run -m pytest
 
 
 [testenv:coverage_report]
+depends =
+    py{311, 310, 39, 38, 37}
 skip_install = true
 deps = coverage[toml]
 commands_pre =