--- a/tox.ini
+++ b/tox.ini
@@ -19,6 +19,8 @@ commands = coverage erase
 
 
 [testenv]
+package = wheel
+wheel_build_env = build_wheel
 depends =
     py{311, 310, 39, 38, 37}: coverage_erase
 extras =