--- a/tox.ini
+++ b/tox.ini
@@ -43,12 +43,20 @@ commands = coverage report
 
 
 [testenv:black]
+# Install the pre-commit environment in the tox environment.
+# This assists with caching in CI.
+setenv =
+    PRE_COMMIT_HOME={envdir}
 skip_install = true
 deps = pre-commit
 commands = pre-commit run --hook-stage manual --all black-check
 
 
 [testenv:isort]
+# Install the pre-commit environment in the tox environment.
+# This assists with caching in CI.
+setenv =
+    PRE_COMMIT_HOME={envdir}
 skip_install = true
 deps = pre-commit
 commands = pre-commit run --hook-stage manual --all isort-check