--- a/tox.ini
+++ b/tox.ini
@@ -22,7 +22,7 @@ commands = coverage erase
 package = wheel
 wheel_build_env = build_wheel
 depends =
-    py{311, 310, 39, 38, 37}: coverage_erase
+    py{311, 310, 39, 38, 37}{-minimum_flask,}: coverage_erase
 extras =
     testing
     !minimum_flask: flask
@@ -33,7 +33,7 @@ commands = coverage run -m pytest
 
 [testenv:coverage_report]
 depends =
-    py{311, 310, 39, 38, 37}
+    py{311, 310, 39, 38, 37}{-minimum_flask,}
 skip_install = true
 deps = coverage[toml]
 commands_pre =
@@ -43,20 +43,20 @@ commands = coverage report
 
 
 [testenv:black]
-# Install the pre-commit environment in the tox environment.
-# This assists with caching in CI.
-setenv =
-    PRE_COMMIT_HOME={envdir}
+# Allow customization of the pre-commit home directory.
+# This helps support caching when tests run in CI.
+passenv =
+    PRE_COMMIT_HOME
 skip_install = true
 deps = pre-commit
 commands = pre-commit run --hook-stage manual --all black-check
 
 
 [testenv:isort]
-# Install the pre-commit environment in the tox environment.
-# This assists with caching in CI.
-setenv =
-    PRE_COMMIT_HOME={envdir}
+# Allow customization of the pre-commit home directory.
+# This helps support caching when tests run in CI.
+passenv =
+    PRE_COMMIT_HOME
 skip_install = true
 deps = pre-commit
 commands = pre-commit run --hook-stage manual --all isort-check