@kurtmckee 2023-01-23
> we could write a workflow file generator!

I'll add Jinja2 as a dev dependency and start coding this up, if you'll write the pre-commit hook to ensure it outputs valid YAML. kthx :stuck_out_tongue_winking_eye: